Update API to process XML file and adjust Docker setup
This commit introduces changes to the API controller to process a XML file instead of a string. The XML file is converted to a JSON string and then sent to a Kafka topic. In addition, this update revises the Docker compose set up removing two volumes related to "mirth-data" and "output-files'. The new dependency "jackson-dataformat-xml" also has been added to the api/pom.xml for XML processing.