Add new features and remove unused ones
Deleted the 'Main' and 'Producteur' classes, and added new features in several modules. The user app was changed significantly with the addition of a graphical user interface and new functionalities in the 'HelloApplication' and 'HelloController' classes. The 'cs1' and 'cs2' modules now have additional functionalities and Dockerfiles. Database initialization was updated with the new 'init-db.sql' file.